requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.4.2
gunicorn==21.2.0